chore: update release workflow for improved security
Change the OSSRH authentication method in the release workflow from
username/password to token-based for better security. Use `OSSRH_TOKEN`
instead of `OSSRH_USERNAME` and `OSSRH_PASSWORD` in the curl command
that notifies the Central Portal. Additionally, modify the commented
namespace for better clarity, and update the notification step to
add verbosity for debugging. The arrow symbol is corrected for
consistency.